feat: save inputs and expected outputs
Now, the test cases (inputs and expected outputs) can be automatically
saved and loaded.

To enable this feature, you have to check the "Save Input and Expected"
option in the "Settings" tab.

This feature doesn't work for temporary buffers. It will save the input
files in the same directory as the saved file, with the case number and
`in`/`ans` suffix. Suppose the saved file is `d:/foo/bar.cpp`, then the
input file for test case 1 will be saved in `d:/foo/bar1.in`, and the
output file for test case 3 will be saved in `d:/foo/bar3.ans`. Empty
inputs/expected outputs won't be saved.

Furthermore, which won't affect users' experience, the codes are
refactored. There used to be lots of duplicate codes for the 3
runners/inputs/outputs/expected outputs, after refactoring, most of them
are replaced by for loops and arrays.

Some minor changes, including fixing some typos, are made in this
commit, too.

Resolves: #25
See also: #52