fix: fix bug when compiling titled unsaved files
Old: Compile titled unsaved files will use the "title" as source file. For example, if the title is `~/a.cpp`, but the editor has no write permission to it, `~/a.cpp` will be passed to the compiler, too. If the title is `~/a`, no suffix will be added to it, and compilation will fail due to the output binary file has the same name as the source file.

New: Compile titled unsaved file will use temporary file, and use "sol.cpp/java/py" as filename.