Adds the EventLogger System
This commit adds a eventLogger which should be used to log any events
from the application as described below :

Step 1 : Include the header "Core/EventLogger.hpp"

Step 2 : Choose the priority order and log the message you want to log.
Each loging function is static and takes two arguments, head and log
body. Keep the head short discription of the log(usually less than 20
chars)

A. Core::Log::i(head, body), it should be used to log informations such as
"Autosave invoked"

B. Core::Log::e(head, body), it should be used to log the Errors that
are not of highest prority. Such as "Python interpreter not found"

C. Core::Log::w(head, body), it should be used to log the warnings, such
as "Run pressed while no testcase exists"

D. Core::Log::wft(head, body), it should be used to log the extremely
critical errors that are unexpected such as unexpected nullptr etc.

These names and shortcuts for looger are derived from Android LogCat
System. It also designates these priorities and takes the logging like
this only.

The logs head has to be single line and body can be any ascii and as
much long as desired. Every time one of the above is invoked, a entry
will be made in the log file, that lies in $(HOME)/cpeditor.log, this
file is temporary and is re-written for each instance of the
application.

Logging does not supports multiple instance of apps as of now.

hotfix(compile)