fix: Set tab stop distance correctly
In the old version, a tab is slightly narrower than <tab width> spaces,
and it's dealed in CP Editor.

After this change, the tab is almost exactly of the same width as
<tab width> spaces, and it's dealed in QCodeEditor.

I guess the reason is that the width of "    " is wider than four times
the width of ' ', and the width of "        " is wider than two times of
the width of "    ", and the delta is smaller than 1 —— it's a floating
number. However, I can't find a function that provides a width in a
floating number.

I googled but found nothing other than the old solution which is not
correct, and I came up with this solution: use the width of
{1000 * <tab width>} spaces divided by 1000.0. It works, and I can't
find a better solution.