Add Detached run for MACOS
This commit fixes #167, Please note that this has been tested and found to be working with C++ and Python language.
The launched detached terminal window that is running the program might not have focus when launched, User should toggle to terminal when running in detached mode.
The newly launched terminal does not vanishes/closes when execution has been done, user need to explicitly close the terminal window.