fix: Set default to empty when setting problem URL if it was empty
In the old version, when setting problem URL, if the
problem URL was not empty, it will be the default, otherwise
"https://codeforces.com/contest//problem" will be the default.

This may be confusing, because one may assume that the default is
always the current problem URL, but there's an exception.

What's more, the URL is long, so it's not so easy to distinguish
"https://codeforces.com/contest//problem" and
"https://codeforces.com/contest/1172/problem/A" in a line edit.

And, in fact, "https://codeforces.com/contest//problem" doesn't make it
much more convenient to set the problem URL.

After this change, the default will always be the current problem URL,
even if it's empty.