fix(preferencewindow): Fix transparency reset to 99 on start if was 100
The bug:

In the old version, if the transparency of the editor is set to 100,
it will be reset to 99 on start.

The reason:

The default maximum value of a QSlider is 99, and the setting is
restored before the maximum value is set to 100. When restoring the
transparency setting, if the setting is 100, it will be set to 99
instead. When setting the value, `QAbstractSlider::valueChanged`
triggers and it changes the setting to 99.

To fix:

Use `void QAbstractSlider::sliderMoved` instead of
`QAbstractSlider::valueChanged`. It triggers on moving the
slider only, and usually doesn't trigger on
`QAbstractSlider::setValue` without pressing the mouse.

("usually" means `setSliderDown(true)` is not called.)