fix: Hide left/right pane in full IO/editor mode
Now the left/right pane will be hidden when changing to full IO/editor
mode. This helps to reduce the width of the window. And the splitter
is not available in full editor/IO mode.

The button "Submit Solution" is also changed to "Submit" for shorter
thus narrower.

The width is important because after adding the "Hide/Show AC" button
it has been wider than half screen on my machine, so splitting the
screen into two parts became hard for me.