fix: Remove delete test cases on disk confirmation
It's strange to delete the test case in the editor but keep it on
the disk, and this dialog can be really annoying, since if you want
to keep it on the disk, it will pop up every time you save the file.
Users can make a back-up by themselves, and there is already
confirmation for delete a test case in the editor.