fix(appwindow): Fix window size issue on Windows
To reproduce on Window 10:

1. Set the transparency to 100 and maximize the window, close it.
2. Open a new session. The window will be still "maximized", but it's
   size is not maximized.

I searched a lot but found nothing similiar to this, so I just used a
workaround: set transparency to 99 before opening tabs and set to 100
later, if the OS is Windows and the transparency is 100.

Also, the default window title in appwindow.ui is set to
"CP Editor: An editor specially designed for competitive programming"
instead of "MainWindow". This is useful when the UI loading is slow.