fix(MainWindow): Fix a bug when running unsaved codes
To reproduce:

1. Open a file.
2. Click Compile and Run (Or press Ctrl+Shift+R).
3. Change the codes before the compilation is finished.
4. The editor will crash when the compilation finishes.

The reason:

Because the code is unsaved, the editor thinks that there was something
when saving the codes, thus it use the temp file instead of the opened
file. When trying to saving the temp file, it tries to save the file
normally at first, it succeeded, so no temp file was created. Then, the
editor thinks the temp file has been created, and makes a crash.

Acting differently when temp file is saved / the opened file is saved
will fix it.