Fix(UpdateNotifier): Fix wrong update version check
Earlier the version compare function did a wrong compare and
reported update available if user had 6.0.1 and github reported
5.9.7, the compare function returned true (update available) but
in reality it is not update. Now we have added check that makes
sure if user version is more than remote latest, the compare
will return false (no update) as expected.