Adds slot and support for language client
Language Server Client is a submodule dependency. This commit adds first
interface and how the client will communicate with the language server.
The editor will fork a server for each instance of app that is open and
communicate with it, all communications are async. For each new tab
opened, please do `client.didOpen()..` and `client.didClose()...`, since
client is shared across all tabs. The result of every response is
arrived in the slots added above. Editor should parse the json returned
in above slots and update the user-interface according.