feat(Find Replace): Enable capturing groups in text to replace
Now in the regex mode of Find/Replace dialog, you can use "\1", "\2"
to refer to the 1st, 2nd capturing group in text to replace.

There is a problem: `\\1` doesn't escape the `\`, but replace the
string by a `\` and the first capturing group. However, it's rare to
replace a string by `\1`, and there's no problem if you don't have any
capturing groups in the text to find.