fix(Update Notifier): Fix version check
When we changed the url of repository the update notifier link was not
changed and the new updates were being looked at old url. Github should
redirect but any re-direction on our UpdateNotifier fails. Now I have
changed the URL to reflect new changes. This is important and must be
release now. I have also added a warning message that will let user know
if github release was failed to fetch not because of internet issue but
because of other issues like the one above.