ci: Set lower cache size for ccache
The cache action has a limit of 5GB, 400M and 600M are too big
and lead to the danger of cache lost.

Big cache size is very useful at local because you may want to
checkout between commits/branches very often. But on CI, it's
not that useful, since the build is usually linear.