feat(Show warning for unsaved settings)
When user closes the preference window without saving the settings a
warning box is shown, user can choose to discard the changes or save
the changes by clicking "Discard" and "Save" button in the warning
messagebox. This partially fixes #189, The part to detect and save
snippets is not implemented, which I believe @ouuan should do most
efficiently because he wrote the snippet system.