feat(TestCases): Add edit input/expected in bigger window; move buttons
Now you can edit input/expected in a bigger window, by the action in
the right-click menu of the textedits.

"Load Input" and "Load Expected" buttons are moved to the right-click
menu of the textedits. "Run" and "Delete" buttons take place of them
(they were in the menu of the "More" button, now the "More" button
is gone).

This closes #190.