Added System wide dark theme for windows 10
This commit fixes #215 and introduces a system wide dark theme that is
based on Fusion style. It can be enabled if the System Theme in Windows
Settings is set to Dark. You will need to restart the application if you
changed system theme while editor was running. You also need to
restart application if in Dark mode you change the Editor theme.