ci: Release on AUR (#222)
* Create release_aur.yml workflow

This workflow releases the latest stable on AUR package as `cpeditor`

* Update release_aur.yml

* For testing purpose remove constraint `on`

* Update release_aur.yml

* Add complete source with submodules for release

The github auto generated sources on release page do not contain
submodules and neither packs .git folder. It is due to this reason that
building latest release after downloading from github release sources,
it not possible. This is important for AUR. Starting this commit any new
release will also include a cpeditor-full-source-x.x.x.tar.gz file that
will contain the complete source code including the submodules.

* Refactor and AUR scripts

This commit adds aur scripts SRCINFO and PKGBUILD to be configured with
CMakeLists.txt for version number. It also refactors the
release_appimage to release_linux

* Test AUR upload #1

* Test AUR upload #2

* Test 3

* Test 4

* Test 5

* Test 6

* Finalize the AUR release workflow

* fix(Missing git config)

* Run aur only for tag push

* Use ubuntu 16.04-LTS and ppa Qt

* Test PPA Qt

* AUR on tag push only

* Final release checks

This commit fixes all the release issues that could appear while releasing the next version here.

* GPL3 to GPLv3+