fix(Preferences): Use default font size and restrict to monospace fonts
In the document of QFont::QFont:

> If pointSize is zero or negative, the point size of the font is set
> to a system-dependent default value. Generally, this is 12 points.

I think it's better to restrict to monospace font before #217 is solved.

In fact, the example in #217, the PragmataPro font, is actually not a
real monospace font, which has ligatures, and doesn't look good on
CP Editor. Users can't have a better experience even if we let them
use that font. What's more, users can modify the settings file manually
in order to set any font.