refactor: Move headers from include to src
If we put the headers and the source files in the same directory, we
can find the files more easily, i.e. don't have to jump between the
include directory and the src directory. Since CP Editor is not a
libary, we don't need to separate the headers and the source files.

The CMakeLists is also sorted.