ci: Add Check Clang Format for PRs
Because of the use of the secrets, the "Auto Clang Format" workflow
only works on push, but not on pull requests. So, if the contributor
didn't enable GitHub Actions in his/her repo, it doesn't run.

However, we still want to know whether a PR meets the format standard
or not, so we can create a new workflow which runs only on PRs, and
only check, not push.