Enhan(Fix Codeacy Issue, better argument parsing)
LSPServer has been marked explicit. It also adds a utility function
named splitArgument, that takes a String and returns a QStringList of
Arguments that can be passed to QProcess. It handles Quotes and Double
quotes nicely and it is based on what PR #216 suggessted. Later we can
use this to parse all other arguments in Compiler command and run
command as well.