fix(LSP): Linting enable and disable
When linting was turned off and editor launched then even after you
enable linting the editor did not linted the current tab rather you
needed to switch to other tab and then lining was done.

After this commit linting of error is done as soon as linting is enabled
from Settings and liniting is cleared and not done as soon as settings
are applied.