fix(default save path): Ignore empty regex and replace captured string (#273)
The bug:

1. Empty regex matches everything.
2. If the regex matches only a part of the problem URL, other parts
   won't be replaced and will be in the saving path.

Now it's fixed.

And the tooltips are modified to point out that the regexes in the
Testcases Matching Rules should match the whole string, and the
regexes in Default File Paths For Problem URLs can match only a part
of the problem URL.

The regexes in the Testcases Matching Rules should match the whole
string, because it's common that a testcase file is a substring of
another.

The regexes in Default File Paths For Problem URLs can match only a
part of the problem URL, because it's nearly impossible that one
problem URL is the substring of another, but sometimes there are
suffixes like "#" or "?" in the problem URL, and users may forget to
include them in the regexes. The user still can enforce whole-string
match by adding "^" at the front of the regex and "$" at the end of
the regex.