feat: Add support for styles and themes in Settings (#404)
* Update submodules

This commits updates the following submodules to latest commits:
	- QCodeEditor: This update fix the issue #232
	- QtFindReplaceDialog: <check commit>
	- singleapplication: <check commit>

* Add changelog & Remove warning for restart

Adds the bug fix to changelog file and removed the warning that
proompted user to restart the application when in dark mode under
windows.

* feat: Support different styles

this commit adds a stylemanager class that should ease the task of
handling the style and themes for the editor. Currently this API is
being used to enable `auto_fusion` for windows OS.

* style: Format codes

* Simplyfy StyleManager and add Light Fusion Style

* style: Format codes

* feat(UI style): add UI style setting and improve style manager

* fix(StyleManager): fix compile error on Windows

* fix(StyleManager): fix broken dark/light fusions

* fix(StyleManager): fix Light Fusion tooltips

* feat(locale): add translation for UI Style

* style(StyleManager): remove wrong namespace comment

* docs(changelog): add UI style settings

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Yufan You <ouuansteve@gmail.com>