Add new signal onEditorTabUpdate(window)
This new signal is emitted by MainWindow to update any UI component of a
Tab or components that are shared across tabs like LSP Instances. This
signal is emitted whenever file is saved to update tab title and it is
also used internally by appwindow to update components upon settings
applied.