feat(Settings): add custom dependency required value
Now the "depends" property is an array of objects, each object
contain "name" and "check", where "name" is the name of a dependency
and "check" is a `std::function<bool(const QVariant &)>` which checks
whether the dependency meets the requirements, the "check" requires
the dependency to be `true` by default. The "check" property in
`settings.json` is not a complete lambda function, it doesn't contain
`[](const QVaraint &var) {` and `}` for convenience.

In this way, you can depend on a setting to be "false", or depend on
non-bool settings.