perf: load MainWindow::EditorStatus without opening an untitled tab
Now you can construct a new MainWindow with a MainWindow::EditorStatus
without opening an untitled tab with a code template at first.

This is about 70% faster when restoring empty files.

This is a part of #191.

This also fixes the bug that the duplicated tab has the same untitled
tab index as the original tab.