feat(Formatter): Add Python formatter YAPF
The clang-formatter is able to format code in C/C++/Java but there was
no formatter support for python language. In this commit we add a new
Extensions YAPF Format that can format any valid python code. It also
accepts many configuration that can be set up by user in style.