ci: use BOT_TOKEN for update-donors (#977)
create-pull-request doesn't trigger other actions when using the default
GITHUB_TOKEN. It's fine for other actions like check-format, because the
targeted base branch is not a protected branch like master. But
update-donors is targeted to the master branch, which needs a personal
token in order to get the required status checks passing.