fix: remove scaling fonts via mouse wheel scroll
This has been reported to be extremely slow, because it actually
applies settings to all tabs besides changing the current editor font.

An option to enable this behavior does not work. Opt-out does not fix
the performance issue. Opt-in does not worth it because the user can
simply use the font setting to change the font size instead. So this
behavior is completely removed.